added "data" service for persitent MobX observable
Service
===
`feathers generate service`
I used mongo to persist because I have a
mongo db running. If you choose another db- you need to re-run the
service command <small>*(and probably `feathers generate {app}`
from the earlier commit...)*</small>

`./src/clientApp.js`
Client side configurations for enabling the service on the client
and providing a means to connect to it

Also, I namespaced my services to `site.com/api/{service}` and increased
the default get to 25 since I don't have anything complex like paging.

MobX
===
All files involved with rendering are going to be accessed on the server
and client, so to make sure the service doesn't try anything funny on
the server we initialize the function <small>(`setupDataStore`)</small>
that integrates the service into the MobX store from the client router.

`setupDataStore` :
- connect event listeners for `[ created , patched , removed ]` to the
store
- set initial state by populating from the service

Note: I used class extension to easily group the store concerns, this
has the side effect of very tight coupling. Probably not the best
way to do this moving forward, but it does allow for modularity with
very a clean form factor, so I like it for now.