added mobx to the server
MobX
===
We started with MobX and a Feathers service already connected and
working on the client. This commit is just to demonstrate how to take
an existing Feathers-[React*-iso*]-MobX build and make it isomorphically
render the MobX data to the server and client.

- added `useStaticRendering` to our store
- returned the service promise to allow the setup function to be
"thenable"
- refactor `setupDataStore` in `router` to use `.then()`
- add `setupDataStore` to `router-server` (as thenable)
- use static if on server, connect service if not:
```js
isServer
    ? useStaticRendering(true)
    : dataService
        .on(//service events etc...
```

That's it! 5 new lines of code across 2 files, and a little refactoring
across 2 files<sup>1</sup>, and we have isomorphic MobX!

<sup>1</sup><sup><small>*(some overlap may apply)*</small></sup>